Fix tests that was passing because a old lense.
Josef Reidinger fixed the ntp lense in this commit:

https://github.com/hercules-team/augeas/commit/01e4c42466b95767d7f761872232a4325ec28b84

Basically before the fix it doesn't expect to have restrict records
with IPv4 explicitly, as for example:

restrict -4 default

So before the fix "-4" was treated as an "address" and "default" as
an option what is wrong.

With the new lense CFA will read the ip version correctly.